+++ lycus/src/Probes/HttpProbe.java
@@ -33,7 +33,7 @@ public class HttpProbe extends BaseProbe {
 	HttpProbe() {
 	}
 
-	public HttpProbe(User user, String probe_id, UUID template_id, String name, long interval, float multiplier,
+	public HttpProbe(User user, String probe_id, UUID template_id, String name, int interval, float multiplier,
 			boolean status, int timeout, String type, String url, String authStatus, String authUsername,
 			String authPassword, boolean deepCheck) {
 		super(user, probe_id, template_id, name, interval, multiplier, status);
@@ -46,7 +46,7 @@ public class HttpProbe extends BaseProbe {
 		this.deepCheck = deepCheck;
 	}
 
-	public HttpProbe(User user, String probe_id, UUID template_id, String name, long interval, float multiplier,
+	public HttpProbe(User user, String probe_id, UUID template_id, String name, int interval, float multiplier,
 			boolean status, int timeout, String type, String url, boolean deepCheck) {
 		super(user, probe_id, template_id, name, interval, multiplier, status);
 		this.httpRequestType = type;
