+++ lycus/src/Probes/NicProbe.java
@@ -44,7 +44,7 @@ public class NicProbe extends BaseProbe {
 	}
 
 	@Override
-	public long getInterval() {
+	public int getInterval() {
 		return discoveryProbe.getElementInterval();
 	}
